Remove copyright notice, template request methods
Semaphores aren't based on the CoreCLR anymore; that was left over
from when I was screwing about trying to get things to work.

Now we can decode JSON using the static methods of entity<T> classes,
and in the future typechecking will be put in to ensure valid json is
passed to these types (right now it will throw a json exception, we
should add custom exception types to make errors easier to debug.

TODO from this point onwards:
- Implement REST API methods in api_client.cpp (see #5)
- Remove `disccord_` prefix from client types
- Add entity types for all discord entities
- Implement websocket API (see #6)
- Implement entity cache
- Clean up includes
- Include a test script (?)
- Expand test range (see #2)
- Write documentation (see #7 for more info)