Add ability to send an array payload
Motivation for this commit is that we need to be able to send up a json array for some endpoints. This started when
I realized that in the current `encode_to` method for models, we could only specify a json object, not just a json array.
So the following overload was added, and a new method `encode_array`, in the model class, so we explicitly know if we expect an object or array.
What I also noticed is how the `request_multi_json` for sending objects is currently handled won't work. there are many
endpoints that only send up a json object, or query strings, that send a json array back. so that was edited. what was added
was another method, `request_multi_json_array` (name can be changed) which denotes we are sending a json array, and expect
a json array back. also `request_json_array` which denotes sending array, expect nothing back.in discords api,
the only endpoints ive found that send an array also receive one, so just this method
was added. its a bit cleaner way as building the array is abstracted to the rest model instead.
One endpoint that required this was implemented to see it put to work, but there is only 1 (maybe 2) more endpoints like this.