Wrapping Guild API (#21)
* add some interface for guild api

* space out the overloads

* should be array of id and pos fields

* impl ep's w/ no json params

* get rid of CRs

* pls tabs

* modify structure, start to add some guild tests

* fix some issues w/ guild_member model

* provide more information when we throw http exceptions, need content length set for non-GET empty resp reqs

* support sending json params on empty requests

* add more tests for guild API

* add more eps

* add more tests

* begin adding more rest models

* Add ability to send an array payload

Motivation for this commit is that we need to be able to send up a json array for some endpoints. This started when
I realized that in the current `encode_to` method for models, we could only specify a json object, not just a json array.
So the following overload was added, and a new method `encode_array`, in the model class, so we explicitly know if we expect an object or array.
What I also noticed is how the `request_multi_json` for sending objects is currently handled won't work. there are many
endpoints that only send up a json object, or query strings, that send a json array back. so that was edited. what was added
was another method, `request_multi_json_array` (name can be changed) which denotes we are sending a json array, and expect
a json array back. also `request_json_array` which denotes sending array, expect nothing back.in discords api,
the only endpoints ive found that send an array also receive one, so just this method
was added. its a bit cleaner way as building the array is abstracted to the rest model instead.
One endpoint that required this was implemented to see it put to work, but there is only 1 (maybe 2) more endpoints like this.

* remove debug print

* wrap more eps

* add test w/ args ctor

* build more rest models

* remove redundant request type