keep new/old values as strings
they will either come in as snowflakes, ints, strings or booleans. originally was going to template them but at the point of decoding we dont actually know which types to supply the template. could do some other magic to make it work, but instead i will just check the json types during decoding and make them all strings. since this info is most likely just going to be logged somewhere, or display in a command, i think its fine to just do this